simple 2.1
更新了图片对比工具和导入sdk工具的功能。前者能够点击图片打开相应的文件夹和选中图片，后者在创建sdk文件夹集合后也能打开文件夹。

+++ Editor/ImportSDKWindow.cs
@@ -53,6 +53,8 @@ public class ImportSDKWindow : EditorWindow {
         if(GUILayout.Button("创建渠道对应的文件夹",GUILayout.Height(30f)))
         {
             CreateAllPlatformDirectory();
+
+            System.Diagnostics.Process.Start(m_path);
         }
 
         EditorGUILayout.BeginVertical(GUI.skin.box);
@@ -172,7 +174,17 @@ public class ImportSDKWindow : EditorWindow {
         //ClearOtherSDK(sdkType);
         try
         {
-            DeleteDirByPath(destPath);
+            if (Directory.Exists(destPath))
+            {
+                DeleteDirByPath(destPath);
+                Debug.Log("testdelete");
+            }
+            else
+            {
+                Debug.Log("testcreate");
+                Directory.CreateDirectory(destPath);
+                AssetDatabase.Refresh();
+            }
 
         CopyFolderTo(sdkPath, destPath , "sdk资源");
